extract theme switcher
